Create precessing_zh_CN.xml
Finished this one :D.
At the first time, I translated all the text and check them in the game. However, the chinese characters seem not to be the okay as the english characters. They were just in a mess and hard to read because of their complex structures, everything is too small to see! So I learn to change the 'size' and 'wrap' (also diable the bold for some odd reasons). And finally I did it, omg.....
Thank u for add the language support for xml, Im really enjoy making localizations (if mDiyo can add the same support that will be very thankful.